+++ src/UmlBuilder.ts
@@ -14,20 +14,28 @@ interface ColumnDataTypeDefaults {
 
 export class UmlBuilder {
 
-	protected flags: TypeormUmlCommandFlags;
 	/**
-	 * Builds database UML and returns it.
+	 * Constructor.
 	 *
 	 * @public
 	 * @param {Connection} connection A database connection.
 	 * @param {TypeormUmlCommandFlags} flags An object with command flags.
+	 */
+	public constructor(
+		protected readonly connection: Connection,
+		protected readonly flags: TypeormUmlCommandFlags
+	) {}
+
+	/**
+	 * Builds database UML and returns it.
+	 *
+	 * @public
 	 * @returns {string} An uml string.
 	 */
-	public buildUml( connection: Connection, flags: TypeormUmlCommandFlags ): string {
-		this.flags = flags;
+	public buildUml(): string {
 		let uml = '@startuml\n\n';
 
-		if ( flags.format === 'txt' ) {
+		if ( this.flags.format === 'txt' ) {
 			uml += '!define pkey(x) x\n';
 			uml += '!define fkey(x) x\n';
 			uml += '!define column(x) x\n';
@@ -36,12 +44,12 @@ export class UmlBuilder {
 			uml += '!define fkey(x) <color:#AAAAAA><&key></color> x\n';
 			uml += '!define column(x) <color:#EFEFEF><&media-record></color> x\n';
 		}
-		uml += `!define table(x) entity x << (T,${ flags.monochrome ? '#FFAAAA' : 'white' }) >>\n\n`;
+		uml += `!define table(x) entity x << (T,${ this.flags.monochrome ? '#FFAAAA' : 'white' }) >>\n\n`;
 
 		uml += 'hide stereotypes\n';
 		uml += 'hide methods\n\n';
 
-		const direction = flags.direction.toUpperCase();
+		const direction = this.flags.direction.toUpperCase();
 		if ( direction === 'LR' ) {
 			uml += 'left to right direction\n';
 		} else if ( direction === 'TB' ) {
@@ -51,8 +59,8 @@ export class UmlBuilder {
 		uml += 'skinparam roundcorner 5\n';
 		uml += 'skinparam linetype ortho\n';
 		uml += 'skinparam shadowing false\n';
-		uml += `skinparam handwritten ${ flags.handwritten ? 'true' : 'false' }\n`;
-		if ( flags.monochrome ) {
+		uml += `skinparam handwritten ${ this.flags.handwritten ? 'true' : 'false' }\n`;
+		if ( this.flags.monochrome ) {
 			uml += 'skinparam monochrome true\n';
 		} else {
 			uml += 'skinparam class {\n';
@@ -62,11 +70,11 @@ export class UmlBuilder {
 			uml += '}\n';
 		}
 
-		const exclude = ( flags.exclude || '' ).split( ',' ).filter( ( item ) => item.trim().length );
-		const include = ( flags.include || '' ).split( ',' ).filter( ( item ) => item.trim().length );
+		const exclude = ( this.flags.exclude || '' ).split( ',' ).filter( ( item ) => item.trim().length );
+		const include = ( this.flags.include || '' ).split( ',' ).filter( ( item ) => item.trim().length );
 
-		const connectionMetadataBuilder = new ConnectionMetadataBuilder( connection );
-		const entityMetadatas = connectionMetadataBuilder.buildEntityMetadatas( connection.options.entities || [] );
+		const connectionMetadataBuilder = new ConnectionMetadataBuilder( this.connection );
+		const entityMetadatas = connectionMetadataBuilder.buildEntityMetadatas( this.connection.options.entities || [] );
 		if ( !entityMetadatas.length ) {
 			throw new Error( 'No entities have been found. Please, check your typeorm config to make sure you have configured it correctly.' );
 		}
@@ -83,7 +91,7 @@ export class UmlBuilder {
 				continue;
 			}
 
-			uml += this.buildClass( entity, connection );
+			uml += this.buildClass( entity );
 			foreignKeys += this.buildForeignKeys( entity );
 		}
 
@@ -101,14 +109,13 @@ export class UmlBuilder {
 	 *
 	 * @protected
 	 * @param {EntityMetadata} entity An entity metadata.
-	 * @param {Connection} connection A database connection.
 	 * @returns {string} An uml class string.
 	 */
-	protected buildClass( entity: EntityMetadata, connection: Connection ): string {
+	protected buildClass( entity: EntityMetadata ): string {
 		let uml = `\ntable( ${ entity.tableNameWithoutPrefix } ) {\n`;
 
 		for ( let i = 0, len = entity.columns.length; i < len; i++ ) {
-			uml += this.buildColumn( entity.columns[i], entity, connection );
+			uml += this.buildColumn( entity.columns[i] );
 		}
 
 		uml += '}\n';
@@ -121,11 +128,9 @@ export class UmlBuilder {
 	 *
 	 * @protected
 	 * @param {ColumnMetadata} column A column metadata.
-	 * @param {EntityMetadata} entity An entity metadata.
-	 * @param {Connection} connection A database connection.
 	 * @returns {string} An uml column string.
 	 */
-	protected buildColumn( column: ColumnMetadata, entity: EntityMetadata, connection: Connection ): string {
+	protected buildColumn( column: ColumnMetadata ): string {
 		let columnName = '';
 		let suffix = '';
 
@@ -139,10 +144,10 @@ export class UmlBuilder {
 		}
 
 		let length = this.getColumnLength( column );
-		const type = connection.driver.normalizeType( column );
+		const type = this.connection.driver.normalizeType( column );
 
-		if ( !length && connection.driver.dataTypeDefaults[type] ) {
-			length = this.getColumnLength( ( connection.driver.dataTypeDefaults[type] as unknown ) as ColumnDataTypeDefaults );
+		if ( !length && this.connection.driver.dataTypeDefaults[type] ) {
+			length = this.getColumnLength( ( this.connection.driver.dataTypeDefaults[type] as unknown ) as ColumnDataTypeDefaults );
 		}
 
 		if ( length ) {
@@ -210,8 +215,8 @@ export class UmlBuilder {
 			return column.precision.toString();
 		}
 
-		if ( ( column as ColumnMetadata ).enum && this.flags['with-enum-values'] ) {
-			return ( column as ColumnMetadata ).enum.join( ',' );
+		if ( this.flags['with-enum-values'] && ( column as ColumnMetadata ).enum ) {
+			return ( column as ColumnMetadata ).enum.join( ', ' );
 		}
 
 		return '';